% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATmv.R
\name{MATmv}
\alias{MATmv}
\title{MATmv: Manejo de datos con casos completos y NA}
\usage{
MATmv(dataframe, ...)
}
\arguments{
\item{dataframe}{Un dataframe que contiene las variables.}

\item{...}{Variables a seleccionar. Pueden ingresarse directamente sin comillas ni vector.}
}
\value{
El nombre del nuevo dataframe con los casos completos.
}
\description{
Esta función selecciona variables de un dataframe, identifica casos completos y crea un resumen
con información sobre los datos faltantes. Si no se especifican variables, se utilizan todas las columnas.
}
\details{
La función genera un nuevo dataframe con los casos completos de las columnas seleccionadas
y una lista de información en el entorno global. La lista contiene una tabla con los datos faltantes
y un gráfico que visualiza los valores faltantes.
}
\examples{
# Usar todas las variables del dataframe:
MATmv(mtcars)

# Usar varias variables sin comillas ni vector:
MATmv(mtcars, mpg, hp, wt)
}
