% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATout.R
\name{MATout}
\alias{MATout}
\title{Identificación de Outliers Univariados}
\usage{
MATout(data, variable)
}
\arguments{
\item{data}{Dataframe que contiene los datos a analizar.}

\item{variable}{Variable continua a analizar (sin comillas).}
}
\value{
La función crea dos objetos en el entorno global:
\itemize{
\item Un nuevo dataframe que contiene los datos sin outliers, nombrado como \code{<nombre_original>_so}.
\item Una lista de información nombrada como \code{<nombre_original>_so_info}, que incluye:
\itemize{
\item \code{Outliers_Table}: Una tabla en formato HTML con los outliers detectados.
\item \code{Boxplot}: Un gráfico de caja de la variable analizada.
}
}
}
\description{
Esta función analiza una variable continua en un dataframe y detecta outliers utilizando el rango intercuartílico (IQR).
Devuelve un dataframe sin los outliers y genera un gráfico de caja y una tabla con los casos identificados como outliers.
}
\examples{
\dontrun{
# Crear un dataframe de ejemplo
datos <- data.frame(altura = c(160, 165, 170, 200, 180, 175, 400))

# Detectar outliers en la variable altura
MATout(datos, altura)
}

}
