% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATout_Mahalanobis.R
\name{MATout_Mahalanobis}
\alias{MATout_Mahalanobis}
\title{Identificación de Outliers Multivariados mediante la Distancia de Mahalanobis}
\usage{
MATout_Mahalanobis(data, ...)
}
\arguments{
\item{data}{Dataframe que contiene los datos a analizar.}

\item{...}{Variables a incluir en el análisis (sin comillas, separadas por comas). Si no se especifican, se procesarán todas las variables numéricas del dataframe.}
}
\value{
La función crea dos objetos en el entorno global:
\itemize{
\item Un nuevo dataframe sin los outliers, nombrado como \code{<nombre_original>_so}.
\item Una lista de información nombrada como \code{<nombre_original>_so_info}, que incluye:
\itemize{
\item \code{Outliers_Table}: Una tabla en formato HTML con los outliers detectados.
\item \code{Boxplot}: Un gráfico de caja de las distancias de Mahalanobis.
}
}
}
\description{
Esta función identifica outliers multivariados utilizando la distancia de Mahalanobis.
Genera un dataframe sin los outliers y proporciona un gráfico de caja y una tabla con los casos detectados como outliers.
}
\examples{
\dontrun{
# Crear un dataframe de ejemplo
datos <- data.frame(x = rnorm(100), y = rnorm(100), z = rnorm(100))

# Detectar outliers multivariados considerando todas las variables numéricas
MATout_Mahalanobis(datos)

# Detectar outliers multivariados considerando solo x e y
MATout_Mahalanobis(datos, x, y)
}

}
